refactor(dashboard): split api/client.ts into domain modules
Split the monolithic 1287-line client.ts into 14 focused domain modules
while maintaining full backward compatibility via re-export hub.

Modules created:
- base.ts (shared infrastructure, fetch wrapper, auth helpers)
- health.ts, updates.ts, metrics.ts, analytics.ts
- sessions.ts, pipelines.ts, auth.ts, sse.ts
- templates.ts, audit.ts, users.ts, session-history.ts, claude.ts

client.ts now re-exports everything from domain modules.
Zero type errors (tsc --noEmit), all 30 client tests pass, build succeeds.
